requirements.txt
python-dotenv==0.14.0
schedule==0.6.0
tweepy==3.8.0
watchdog==0.10.3